+++ CHANGELOG.md
@@ -7,7 +7,10 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
 * Connectors
    * Show Managed connectors
    * Error scenarios implemented in Table component
-
+* News Section
+   * Fixed UI of news section in home page
+* Sort section
+   * Hover effect added
 ## 1.4.0
 
 ### Change